Intermediate: Use TLS approach to prevent reentrancy.
Adding a new TLS based approach to detect and prevent reentrancy that
results in stack overflow. In the patch handler, we read a TLS value and
if that value is set, we bail out. If the value is not set, then we set
it to prevent re-entrancy in the same function.