Support custom sprintf and remove msvc from exclusion.
This checkin significantly improves the reliability of tracing by
switching to a custom routing (and not calling into CRT). It also
removed msvc from hardcoded exclusion because there is no crash by
hooking msvc now. Other exclusions are still left but they may be
removed at a later time.

When tested with removal of ntdll, notepad works fine but the dialog
that opens the file doesn't paint correctly. Similar issues might exist
for other exclusions so they are left in as it is for now.