IM: Shared memory based trace is working.
We still need to print trace output properly. One other important issue
remaining is related to killing stracent, which leaves the DLL loaded in
the victim process. The DLL would run out of trace buffers, once
straceNT is killed, thereby causing patched functions in victim to spin
forever.

One potential solution is to make the buffers non-blocking or switch
trace output to OutputDebugString. The issue with that approach is that
if straceNT is force killed, then we can't do much. We need to think of
a better solution to this problem.